+++ packages/starter/lib/packageJSONPreparing.js
@@ -19,7 +19,7 @@ const packageJSONPreparing = async (packageJSON, {
           { name: 'react-dom', version: '16' },
         ],
         devDependencies: [
-          { name: '@rockpack/compiler', version: '1.0.0' }
+          { name: '@rockpack/compiler', version: '1.1.0' }
         ]
       });
 
@@ -50,12 +50,12 @@ const packageJSONPreparing = async (packageJSON, {
           { name: 'isomorphic-style-loader', version: '5.1.0' },
           { name: 'pretty-error', version: '2.1.1' },
           { name: '@koa/router', version: '8' },
-          { name: '@rockpack/ussr', version: '1.0.0' },
+          { name: '@rockpack/ussr', version: '1.1.0' },
           { name: '@loadable/component', version: '5.13.1' },
           { name: '@loadable/server', version: '5.13.1' }
         ],
         devDependencies: [
-          { name: '@rockpack/compiler', version: '1.0.0' }
+          { name: '@rockpack/compiler', version: '1.1.0' }
         ]
       });
 
@@ -100,12 +100,12 @@ const packageJSONPreparing = async (packageJSON, {
           { name: 'isomorphic-style-loader', version: '5.1.0' },
           { name: 'pretty-error', version: '2.1.1' },
           { name: '@koa/router', version: '8' },
-          { name: '@rockpack/ussr', version: '1.0.0' },
+          { name: '@rockpack/ussr', version: '1.1.0' },
           { name: '@loadable/component', version: '5.13.1' },
           { name: '@loadable/server', version: '5.13.1' }
         ],
         devDependencies: [
-          { name: '@rockpack/compiler', version: '1.0.0' }
+          { name: '@rockpack/compiler', version: '1.1.0' }
         ]
       });
 
@@ -131,7 +131,7 @@ const packageJSONPreparing = async (packageJSON, {
     case 'library':
       packageJSON = await addDependencies(packageJSON, {
         devDependencies: [
-          { name: '@rockpack/compiler', version: '1.0.0' }
+          { name: '@rockpack/compiler', version: '1.1.0' }
         ]
       });
       packageJSON = addFields(packageJSON, {
@@ -154,7 +154,7 @@ const packageJSONPreparing = async (packageJSON, {
     case 'nodejs':
       packageJSON = await addDependencies(packageJSON, {
         devDependencies: [
-          { name: '@rockpack/compiler', version: '1.0.0' }
+          { name: '@rockpack/compiler', version: '1.1.0' }
         ]
       });
 
@@ -163,7 +163,7 @@ const packageJSONPreparing = async (packageJSON, {
           dependencies: [],
           devDependencies: [
             { name: '@types/node', version: '14' },
-            { name: '@rockpack/compiler', version: '1.0.0' }
+            { name: '@rockpack/compiler', version: '1.1.0' }
           ]
         });
       }
@@ -188,7 +188,7 @@ const packageJSONPreparing = async (packageJSON, {
     });
     packageJSON = await addDependencies(packageJSON, {
       devDependencies: [
-        { name: '@rockpack/codestyle', version: '1.0.0' }
+        { name: '@rockpack/codestyle', version: '1.1.0' }
       ]
     });
   }
@@ -199,7 +199,7 @@ const packageJSONPreparing = async (packageJSON, {
     });
     packageJSON = await addDependencies(packageJSON, {
       devDependencies: [
-        { name: '@rockpack/tester', version: '1.0.0' }
+        { name: '@rockpack/tester', version: '1.1.0' }
       ]
     });
     if (appType === 'csr' || appType === 'ssr-light' || appType === 'ssr-full') {
@@ -221,14 +221,14 @@ const packageJSONPreparing = async (packageJSON, {
     });
     packageJSON = await addDependencies(packageJSON, {
       dependencies: [
-        { name: '@rockpack/localazer', version: '1.0.0' }
+        { name: '@rockpack/localazer', version: '1.1.0' }
       ]
     });
   }
   if (modules.logger) {
     packageJSON = await addDependencies(packageJSON, {
       dependencies: [
-        { name: '@rockpack/logger', version: '1.0.0' }
+        { name: '@rockpack/logger', version: '1.1.0' }
       ]
     });
   }