+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # Changelog
 
+## [1.1.0]
+- fix(@rockpack/compiler): Side effects
+- Small fixes
+- added: Backbone for E2E tests
+
+BREAKING CHANGES:
+
+- Simplification of USSR API
+- .modules.(s)css/less -> .module.(s)css/less (create-react-app compatibility)
+
 ## [1.0.0]